+++ common/geography_utils/src/height.cpp
@@ -21,17 +21,19 @@
 #include <string>
 #include <utility>
 
-namespace tier4_geography_utils
+namespace geography_utils
 {
 
-double convert_wgs84_to_egm2008(const double height, const double latitude, const double longitude)
+double convert_wgs84_to_egm2008(
+  const double height, const double latitude, const double longitude)
 {
   GeographicLib::Geoid egm2008("egm2008-1");
   // cSpell: ignore ELLIPSOIDTOGEOID
   return egm2008.ConvertHeight(latitude, longitude, height, GeographicLib::Geoid::ELLIPSOIDTOGEOID);
 }
 
-double convert_egm2008_to_wgs84(const double height, const double latitude, const double longitude)
+double convert_egm2008_to_wgs84(
+  const double height, const double latitude, const double longitude)
 {
   GeographicLib::Geoid egm2008("egm2008-1");
   // cSpell: ignore GEOIDTOELLIPSOID
@@ -61,4 +63,4 @@ double convert_height(
   }
 }
 
-}  // namespace tier4_geography_utils
+}  // namespace geography_utils
